add -e flag to pip install command
For some reason setuptools only seems to recognise srim as a package when EVA is build with the -e flag. I have tried to fix this for days but at this point I think it's easier to just encourage users to install it as editable as there isnt really any downside to it other than it being a little slower.